Revert "refactor: move and rename hooks.ts"
This reverts commit 94dbb760e6d6218748ed4ae263d85d099ba775ff.
